refactor(web): extract ManualExpenseSheet category system to fix max-lines gate
ManualExpenseSheet.tsx was 601 effective LOC — over the 600 `max-lines` gate
(Hard Rule #18), so `pnpm lint` was red on it (the "red lint gate" flagged by
the 2026-05-29 drift audit; my earlier 0013 closure note "allowlist empty /
lint green" was inaccurate — an empty allowlist doesn't mean every file is
under the threshold).

Extract the self-contained category-slug system (types, CATEGORY_DISPLAY,
the three-era upgradeCategory + legacy maps) into a new pure module
`manualExpenseCategories.ts`. The component imports what it needs and
re-exports `upgradeCategory` / `CATEGORY_DISPLAY` / `CategorySlug` for
backward-compat (the test imports `upgradeCategory`). No behaviour change.
ManualExpenseSheet.tsx now passes `max-lines`.
